Rename header env_var_utils.h to parse_env_var.h
