+++ onnxruntime/test/util/test_random_seed.cc
@@ -5,7 +5,7 @@
 
 #include <chrono>
 
-#include "core/platform/env_var_utils.h"
+#include "core/platform/parse_env_var.h"
 
 namespace onnxruntime {
 namespace test {
